+++ src/SampleCode.csproj
@@ -96,14 +96,14 @@
     <Compile Include="Samples\Authentication\GenerateHttpRequestHeaders\GetGenerateHeaders.cs" />
     <Compile Include="Samples\Authentication\GenerateHttpRequestHeaders\PostGenerateHeaders.cs" />
     <Compile Include="Samples\Authentication\GenerateHttpRequestHeaders\PutGenerateHeaders.cs" />
-    <Compile Include="Samples\Decision_Manager\CoreServices\CreateDecisionManagerRequest.cs" />
-    <Compile Include="Samples\Decision_Manager\CoreServices\DecisionProfileReject.cs" />
-    <Compile Include="Samples\Decision_Manager\CoreServices\DmWithBuyerInformation.cs" />
-    <Compile Include="Samples\Decision_Manager\CoreServices\DmWithDeviceInformation.cs" />
-    <Compile Include="Samples\Decision_Manager\CoreServices\DmWithMerchantDefinedInformation.cs" />
-    <Compile Include="Samples\Decision_Manager\CoreServices\DmWithShippingInformation.cs" />
-    <Compile Include="Samples\Decision_Manager\CoreServices\DmWithTravelInformation.cs" />
-    <Compile Include="Samples\Decision_Manager\CoreServices\ScoreExceedsThreshold.cs" />
+    <Compile Include="Samples\DecisionManager\CoreServices\CreateDecisionManagerRequest.cs" />
+    <Compile Include="Samples\DecisionManager\CoreServices\DecisionProfileReject.cs" />
+    <Compile Include="Samples\DecisionManager\CoreServices\DmWithBuyerInformation.cs" />
+    <Compile Include="Samples\DecisionManager\CoreServices\DmWithDeviceInformation.cs" />
+    <Compile Include="Samples\DecisionManager\CoreServices\DmWithMerchantDefinedInformation.cs" />
+    <Compile Include="Samples\DecisionManager\CoreServices\DmWithShippingInformation.cs" />
+    <Compile Include="Samples\DecisionManager\CoreServices\DmWithTravelInformation.cs" />
+    <Compile Include="Samples\DecisionManager\CoreServices\ScoreExceedsThreshold.cs" />
     <Compile Include="Samples\Flex\CoreServices\GenerateKey.cs" />
     <Compile Include="Configuration.cs" />
     <Compile Include="Samples\Flex\CoreServices\TokenizeCard.cs" />
