+++ src/Samples/DecisionManager/CoreServices/DMwithMerchantDefinedInformation.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 using CyberSource.Api;
 using CyberSource.Model;
 
-namespace Cybersource_rest_samples_dotnet.Samples.Decision_Manager.CoreServices
+namespace Cybersource_rest_samples_dotnet.Samples.DecisionManager.CoreServices
 {
     public class DmWithMerchantDefinedInformation
     {
